tweaking CSR controller to hide some OCM specifics
